+++ CHANGELOG.rst
@@ -6,9 +6,11 @@ Changelog
 
 - Updated to hikari (up-to-date discord python lib)
 - Fix timer update when finishing (display as finished, not with 1 or 2 seconds left)
-- Add "secured" option so that only the owner can modify the timer
-- Tiny improvements on the timer logic (more precise finish)
+- Add "protected" option so that only the owner can modify the timer
+- Tiny improvements on the timer logic (more precise countdown and finish)
+- Avoid double displays in some corner cases (hanging timer besides a running one)
 - Graceful shutdown (all timers stop rather than being left hanging)
+- Overall performance improvements (uvloop, hikari speedups, optimized python)
 
 1.3 (2022-12-01)
 ----------------